Revert "Change schemas to better match TPC-H and airlift"
This reverts commit 0f988950f68281b8fd27bfa081dc6f97b3d8a3d0.

+++ tempto-core/src/test/groovy/com/teradata/tempto/internal/fulfillment/hive/HiveTableManagerTest.groovy
@@ -154,7 +154,7 @@ class HiveTableManagerTest
 
   private static final String NATION_DDL_TEMPLATE = '''
     CREATE TABLE %NAME%(
-            n_nationid BIGINT,
+            n_nationid INT,
             n_name STRING)
             ROW FORMAT DELIMITED FIELDS TERMINATED BY '|'
             LOCATION '%LOCATION%'
@@ -171,7 +171,7 @@ class HiveTableManagerTest
 
   private static final String PARTITIONED_NATION_DDL_TEMPLATE = '''
     CREATE TABLE %NAME%(
-            n_nationid BIGINT,
+            n_nationid INT,
             n_name STRING)
             ROW FORMAT DELIMITED FIELDS TERMINATED BY '|'
 '''
